SignedDistanceToMeshOptions
