+++ source/MRMesh/MRMesh.vcxproj.filters
@@ -1248,6 +1248,9 @@
     <ClInclude Include="MRMeshTopologyDiff.h">
       <Filter>Source Files\Mesh</Filter>
     </ClInclude>
+    <ClInclude Include="MRDistanceToMeshOptions.h">
+      <Filter>Source Files\AABBTree</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="MRParallelProgressReporter.cpp">
